Resolve nsenter fork exception
with the consolodation of container attachment methods we've seen issues
where attaching to a container a container can result in the following
error.

> __libc_fork: Assertion `THREAD_GETMEM (self, tid) != ppid'

This error is caused when two container enter commands are executed at
the exact same time and the OS is not able to accomodate the
simultanious fork. To resolve the issue a retry decorator has been added
which will re-execute the failed command within a one third second
delay.

Change-Id: I97cb3f92ae67b2c5c098c8792a0793972ec0105a
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>